requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
torch>=2.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
